[bug #161336671]Code refactor
